added OS X project files
most of these examples ran on OS X 10.10 with OF 0.8.4.
but everything that contains #include <array> did not compile.
also these examples did not have any visual output.
NOC8_9__Simple_L_system_sent
NOC9_1__Genetic_algorithm__Evo
NOC10_3__Neural_network_diagr

+++ Nature Of Code OF/NOC7_1__Wolfram_elementary_c/src/CA.cpp
@@ -59,10 +59,11 @@ int CA::rules(int a, int b, int c)
     sstm << a << b << c;
     string result = sstm.str();
     int decimal = 0;
-    for(int counter = 0; counter < result.size(); counter++)
+    for(double counter = 0; counter < result.size(); counter++)
     {
-        if(result.c_str()[counter] == '1')
+        if(result.c_str()[(int)counter] == '1')
             decimal += pow(2, counter);
+   
     }
 
     // std::cout << "Rules value: " << a << b << c << " Decimal: " << decimal << ".\n";
